requirements.txt
# AI Gamestore Harness - dependencies
playwright>=1.40.0
Pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# LLM providers (install only the ones you use)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.8.0
together>=1.0.0
